test that the reader context manager closes open files; test that the shapefile type and length are correct
